+++ .idea/dataSources.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="DataSourceManagerImpl" format="xml" multifile-model="true">
-    <data-source source="LOCAL" name="Homework15@localhost" uuid="c22c1e7e-6f20-470a-852f-8011963dce0a">
+    <data-source source="LOCAL" name="Homework14@localhost" uuid="c22c1e7e-6f20-470a-852f-8011963dce0a">
       <driver-ref>postgresql</driver-ref>
       <synchronize>true</synchronize>
       <jdbc-driver>org.postgresql.Driver</jdbc-driver>
