+++ homework-15-rest-service-on-spring/src/main/java/ru/otus/java/professional/yampolskiy/spring/homework15restserviceonspring/Homework15RestServiceOnSpringApplication.java
@@ -7,15 +7,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Homework15RestServiceOnSpringApplication implements CommandLineRunner {
+public class Homework15RestServiceOnSpringApplication{
 
     private static final Logger logger = LoggerFactory.getLogger(Homework15RestServiceOnSpringApplication.class.getName());
 
     public static void main(String[] args) {
         SpringApplication.run(Homework15RestServiceOnSpringApplication.class, args);
-    }
-
-    @Override
-    public void run(String... args) throws Exception {
     }
 }
